requirements.txt
pyyaml>=6.0
matplotlib>=3.7.0
networkx>=3.1
tabulate>=0.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
colorama>=0.4.6